added rational() in rational.py
This function is not a method of Rational class.
It converts a int n to Rational(n,1),
and Rational is left as it is.

So if you want a result of Rational(2,3) / Rational(1,3) in Rational class
(i.e. Rational(2,1)),
you may write "rational(Rational(2,3) / Rational(1,3))",
and if you want in int value (i.e. 2),
you may write simply "Rational(2,3) / Rational(1,3))".